services/telemetry-stream/pyproject.toml
[project]
name = "telemetry-stream"
version = "0.1.0"
description = "telemetry-stream service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116,<1",
  "uvicorn[standard]>=0.35,<1",
  "pydantic>=2.8,<3",
  "event-contracts>=0.1.0",
  "nats-py>=2.8,<3",
  "prometheus-client>=0.23,<1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
